requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
slowapi==0.1.9
requests==2.31.0
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
numpy==1.24.3
aiofiles==23.2.1
python-magic==0.4.27
